+++ python/GafferSceneUI/SceneEditor.py
@@ -65,6 +65,19 @@ class SceneEditor( GafferUI.NodeSetEditor ) :
 
 		self.__parentingConnections = {}
 
+		self.__globalEditTargetLinked = False
+		self.parentChangedSignal().connect( Gaffer.WeakMethod( self.__parentChanged ) )
+
+	def editScope( self ) :
+
+		if not "editScope" in self.settings() :
+			return None
+
+		return Gaffer.PlugAlgo.findSource(
+			self.settings()["editScope"],
+			lambda plug : plug.node() if isinstance( plug.node(), Gaffer.EditScope ) else None
+		)
+
 	def _updateFromSet( self ) :
 
 		# Find ScenePlugs and connect them to `settings()["in"]`.
@@ -123,6 +136,16 @@ class SceneEditor( GafferUI.NodeSetEditor ) :
 			_reverseNodes = True,
 			_ellipsis = False
 		)
+
+	def __parentChanged( self, widget ) :
+
+		if self.__globalEditTargetLinked or not "editScope" in self.settings() :
+			return
+
+		compoundEditor = self.ancestor( GafferUI.CompoundEditor )
+		if compoundEditor :
+			self.settings()["editScope"].setInput( compoundEditor.settings()["editScope"] )
+			self.__globalEditTargetLinked = True
 
 	def __scenePlugParentChanged( self, plug, newParent ) :