+++ python/GafferUI/ScriptWindow.py
@@ -37,6 +37,8 @@
 
 import weakref
 
+import imath
+
 import IECore
 
 import Gaffer
@@ -56,8 +58,14 @@ class ScriptWindow( GafferUI.Window ) :
 
 		self.__listContainer = GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Vertical, spacing = 0 )
 
+		self.__menuContainer = GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 )
+		self.__listContainer.append( self.__menuContainer )
+
 		menuDefinition = self.menuDefinition( script.applicationRoot() ) if script.applicationRoot() else IECore.MenuDefinition()
-		self.__listContainer.append( GafferUI.MenuBar( menuDefinition ) )
+		self.__menuBar = GafferUI.MenuBar( menuDefinition )
+		self.__menuBar.addShortcutTarget( self )
+		self.__menuContainer.append( self.__menuBar )
+		self.__menuContainer._qtWidget().setObjectName( "gafferMenuBarWidgetContainer" )
 		# Must parent `__listContainer` to the window before setting the layout,
 		# because `CompoundEditor.__parentChanged` needs to find the ancestor
 		# ScriptWindow.
@@ -76,7 +84,7 @@ class ScriptWindow( GafferUI.Window ) :
 
 	def menuBar( self ) :
 
-		return self.__listContainer[0]
+		return self.__menuContainer[0]
 
 	def scriptNode( self ) :
 
@@ -84,11 +92,29 @@ class ScriptWindow( GafferUI.Window ) :
 
 	def setLayout( self, compoundEditor ) :
 
+		# When changing layout we need to manually transfer the edit scope
+		# from an existing CompoundEditor to the new one.
+		currentEditScope = None
 		if len( self.__listContainer ) > 1 :
+			currentEditScope = self.getLayout().settings()["editScope"].getInput()
 			del self.__listContainer[1]
 
 		assert( compoundEditor.scriptNode().isSame( self.scriptNode() ) )
 		self.__listContainer.append( compoundEditor, expand=True )
+
+		if len( self.__menuContainer ) > 1 :
+			del self.__menuContainer[1:]
+
+		if currentEditScope is not None :
+			compoundEditor.settings()["editScope"].setInput( currentEditScope )
+		self.__menuContainer.append(
+			GafferUI.PlugLayout(
+				compoundEditor.settings(),
+				orientation = GafferUI.ListContainer.Orientation.Horizontal,
+				rootSection = "Settings"
+			)
+		)
+		self.__menuContainer.append( GafferUI.Spacer( imath.V2i( 0 ) ) )
 
 	def getLayout( self ) :